Move `balance_change_notifier` to `controller`
