+++ integration_test/wallet_test.dart
@@ -8,7 +8,7 @@ import 'package:ten_ten_one/balance.dart';
 import 'package:ten_ten_one/wallet/wallet_dashboard.dart';
 import 'package:ten_ten_one/wallet/wallet_lightning.dart';
 import 'package:ten_ten_one/model/amount.dart';
-import 'package:ten_ten_one/model/balance_change_notifier.dart';
+import 'package:ten_ten_one/controller/balance_change_notifier.dart';
 import 'package:ten_ten_one/model/seed_backup_change_notifier.dart';
 import 'package:ten_ten_one/wallet/seed.dart';
